Creare  package for symbolic value classes
