+++ javascript-frontend/src/main/java/org/sonar/javascript/se/sv/package-info.java
@@ -17,12 +17,5 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonar.javascript.se;
-
-import java.util.List;
-
-interface SymbolicValue {
-
-  List<ProgramState> constrain(ProgramState state, Constraint constraint);
-
-}
+@javax.annotation.ParametersAreNonnullByDefault
+package org.sonar.javascript.se.sv;
