+++ javascript-frontend/src/test/java/org/sonar/javascript/se/SymbolicValueTest.java
@@ -20,6 +20,7 @@
 package org.sonar.javascript.se;
 
 import org.junit.Test;
+import org.sonar.javascript.se.sv.SymbolicValue;
 import org.sonar.plugins.javascript.api.symbols.Symbol;
 
 import static org.fest.assertions.Assertions.assertThat;
